requirements.txt
feedparser==6.0.11
requests==2.31.0
python-dotenv==1.0.0
supabase==2.3.0
rich==13.7.0
httpx==0.24.1